requirements.txt
streamlit
requests
feedparser
beautifulsoup4
sentence-transformers>=2.2.0
transformers>=4.35.0
spacy>=3.7.0
scikit-learn
nltk
uvicorn
python-dotenv
schedule
torch>=2.0.0
torchaudio
torchvision
newsapi>=1.0
pydantic-settings
aiohttp
PyJWT
